simplify pos

+++ src/team.hpp
@@ -1,7 +1,9 @@
 #ifndef TEAM_HPP
 #define TEAM_HPP
 
-enum class Team : int { White = 0, Black = 1 };
+#include <cstdint>
+
+enum class Team : uint8_t { White = 0, Black = 1 };
 inline Team oppositeTeam(Team team) {
     return (team == Team::White) ? Team::Black : Team::White;
 };
